1.0 ? remake build.gradle add essential Make Ignore Overlay settings apply instantly Add the option to retain overlays while randomly selecting a pack
