+++ build.gradle
@@ -1,19 +1,8 @@
-buildscript {
-    repositories {
-        jcenter()
-        maven {
-            name = "forge"
-            url = "http://files.minecraftforge.net/maven"
-        }
-
-        flatDir dirs: 'libs'
-    }
-    dependencies {
-        classpath 'net.minecraftforge.gradle:ForgeGradle:2.1-SNAPSHOT'
-    }
+plugins {
+    id "java"
+    id "com.github.johnrengelman.shadow" version "6.1.0"
+    id "net.minecraftforge.gradle.forge" version "86b2392"
 }
-
-apply plugin: 'net.minecraftforge.gradle.forge'
 
 version = project.modVersion
 group= project.modGroup
@@ -26,11 +15,25 @@ minecraft {
     runDir = "run"
     mappings = project.mcpVersion
     makeObfSourceJar = false
+    clientRunArgs += '--tweakClass gg.essential.loader.stage0.EssentialSetupTweaker'
+}
+
+configurations {
+    // Creates an extra configuration that implements `implementation` to be used later as the configuration that shades libraries
+    include
+    implementation.extendsFrom(include)
+}
+
+repositories {
+    mavenCentral()
+    maven { url 'https://repo.sk1er.club/repository/maven-releases/' }
+    maven { url 'https://jitpack.io/' }
+    maven { url 'https://repo.spongepowered.org/repository/maven-public/' }
 }
 
 dependencies {
-    // include modcore as a dependency
-    provided files("libs/modcore.jar")
+    include ('gg.essential:loader-launchwrapper:1.1.0')
+    provided 'gg.essential:essential-1.8.9-forge:1328'
 }
 
 processResources {
@@ -54,8 +57,40 @@ processResources {
 
 jar {
     manifest {
-        attributes "FMLCorePlugin": "dev.fxe.mods.resourcepackdisplay.modcore.forge.FMLLoadingPlugin",
+        attributes 'TweakClass': 'gg.essential.loader.stage0.EssentialSetupTweaker',
                 "ModSide": "CLIENT",
-                "FMLCorePluginContainsFMLMod": true
+                'TweakOrder': '0'
     }
+}
+
+
+// This forces the build task to run shadowJar after
+tasks.reobfJar.dependsOn(tasks.shadowJar)
+
+// This adds support to ("embed", "shade", "include") libraries into our JAR
+shadowJar {
+    archiveClassifier.set("")
+    configurations = [project.configurations.include]
+    duplicatesStrategy DuplicatesStrategy.EXCLUDE
+
+
+    exclude 'LICENSE.md'
+    exclude 'pack.mcmeta'
+    exclude 'dummyThing'
+    exclude '**/module-info.class'
+    exclude '*.so'
+    exclude '*.dylib'
+    exclude '*.dll'
+    exclude '*.jnilib'
+    exclude 'ibxm/**'
+    exclude 'com/jcraft/**'
+    exclude 'org/lwjgl/**'
+    exclude 'net/java/**'
+
+    exclude 'META-INF/proguard/**'
+    exclude 'META-INF/maven/**'
+    exclude 'META-INF/versions/**'
+    exclude 'META-INF/com.android.tools/**'
+
+    exclude 'fabric.mod.json'
 }
