+++ gradle.properties
@@ -1,9 +1,8 @@
 modGroup=dev.fxe
-modVersion=1.4-Beta
+modVersion=1.0
 modBaseName=ResourcePackDisplay
 forgeVersion=1.8.9-11.15.1.2318-1.8.9
 mcpVersion=stable_22
-tweakClass=dev.fxe.mods.resourcepackdisplay.tweaker.TweakClass
 
 org.gradle.daemon=true
 org.gradle.parallel=true
