+++ src/main/java/dev/fxe/mods/resourcepackdisplay/ui/GUI.java
@@ -45,7 +45,7 @@ public class GUI extends GuiScreen {
     @Override
     public void drawScreen(int mouseX, int mouseY, float partialTicks) {
         updatePos(mouseX, mouseY);
-        UI.drawPackDisplay();
+        HUD.drawPackDisplay();
         super.drawScreen(mouseX, mouseY, partialTicks);
     }
